refactor detected event code
