+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Refactor the event processing code by separating it into protocol files.
+
 ## [0.14.2] - 2023-08-22
 
 ### Added
@@ -271,6 +277,7 @@ across our system.
 
 - An initial version.
 
+[Unreleased]: https://github.com/petabi/review-web/compare/0.14.1...main
 [0.14.2]: https://github.com/petabi/review-web/compare/0.14.1...0.14.2
 [0.14.1]: https://github.com/petabi/review-web/compare/0.14.0...0.14.1
 [0.14.0]: https://github.com/petabi/review-web/compare/0.13.1...0.14.0
