+++ src/graphql/event.rs
@@ -1,6 +1,19 @@
+mod conn;
+mod dns;
+mod ftp;
 mod group;
+mod http;
+mod ldap;
+mod rdp;
 
 pub(super) use self::group::EventGroupQuery;
+use self::{
+    conn::BlockListConn, conn::ExternalDdos, conn::MultiHostPortScan, conn::PortScan,
+    dns::BlockListDns, dns::CryptocurrencyMiningPool, dns::DnsCovertChannel, ftp::FtpBruteForce,
+    ftp::FtpPlainText, http::DomainGenerationAlgorithm, http::HttpThreat, http::NonBrowser,
+    http::RepeatedHttpSessions, http::TorConnection, ldap::LdapBruteForce, ldap::LdapPlainText,
+    rdp::RdpBruteForce,
+};
 use super::{
     customer::{Customer, HostNetworkGroupInput},
     filter::{FlowKind, LearningMethod, TrafficDirection},
@@ -14,7 +27,6 @@ use async_graphql::{
 };
 use bincode::Options;
 use chrono::{DateTime, Utc};
-use database::EventKind;
 use futures::channel::mpsc::{unbounded, UnboundedSender};
 use futures_util::stream::Stream;
 use num_traits::FromPrimitive;
@@ -23,7 +35,8 @@ use review_database::{
     event::RecordType,
     find_ip_country,
     types::{Endpoint, EventCategory, FromKeyValue, HostNetworkGroup},
-    Direction, EventFilter, EventIterator, IndexedMap, IndexedMultimap, IterableMap, Store,
+    Direction, EventFilter, EventIterator, EventKind, IndexedMap, IndexedMultimap, IterableMap,
+    Store,
 };
 use std::{
     cmp,
@@ -445,148 +458,6 @@ fn country_code(ctx: &Context<'_>, addr: IpAddr) -> String {
     }
 }
 
-struct DnsCovertChannel {
-    inner: database::DnsCovertChannel,
-}
-
-#[Object]
-impl DnsCovertChannel {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn session_end_time(&self) -> DateTime<Utc> {
-        self.inner.session_end_time
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn query(&self) -> &str {
-        &self.inner.query
-    }
-
-    async fn answer(&self) -> Vec<String> {
-        self.inner.answer.clone()
-    }
-
-    async fn trans_id(&self) -> u16 {
-        self.inner.trans_id
-    }
-
-    async fn rtt(&self) -> i64 {
-        self.inner.rtt
-    }
-
-    async fn qclass(&self) -> u16 {
-        self.inner.qclass
-    }
-
-    async fn qtype(&self) -> u16 {
-        self.inner.qtype
-    }
-
-    async fn rcode(&self) -> u16 {
-        self.inner.rcode
-    }
-
-    async fn aa_flag(&self) -> bool {
-        self.inner.aa_flag
-    }
-
-    async fn tc_flag(&self) -> bool {
-        self.inner.tc_flag
-    }
-
-    async fn rd_flag(&self) -> bool {
-        self.inner.rd_flag
-    }
-
-    async fn ra_flag(&self) -> bool {
-        self.inner.ra_flag
-    }
-
-    async fn ttl(&self) -> Vec<i32> {
-        self.inner.ttl.clone()
-    }
-
-    async fn confidence(&self) -> f32 {
-        self.inner.confidence
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::DnsCovertChannel> for DnsCovertChannel {
-    fn from(inner: database::DnsCovertChannel) -> Self {
-        Self { inner }
-    }
-}
-
 fn find_ip_customer(map: &IndexedMap, addr: IpAddr) -> Result<Option<Customer>> {
     for (key, value) in map.iter_forward()? {
         let customer = database::Customer::from_key_value(key.as_ref(), value.as_ref())?;
@@ -605,1929 +476,6 @@ fn find_ip_network(map: &IndexedMultimap, addr: IpAddr) -> Result<Option<Network
         }
     }
     Ok(None)
-}
-
-struct HttpThreat {
-    inner: database::HttpThreat,
-}
-
-#[Object]
-impl HttpThreat {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn method(&self) -> &str {
-        &self.inner.method
-    }
-
-    async fn host(&self) -> &str {
-        &self.inner.host
-    }
-
-    async fn content(&self) -> String {
-        format!(
-            "{} {} {} {} {} {}",
-            self.inner.method,
-            self.inner.host,
-            self.inner.uri,
-            self.inner.referer,
-            self.inner.status_code,
-            self.inner.user_agent
-        )
-    }
-
-    async fn uri(&self) -> &str {
-        &self.inner.uri
-    }
-
-    async fn referer(&self) -> &str {
-        &self.inner.referer
-    }
-
-    async fn version(&self) -> &str {
-        &self.inner.version
-    }
-
-    async fn user_agent(&self) -> &str {
-        &self.inner.user_agent
-    }
-
-    async fn request_len(&self) -> usize {
-        self.inner.request_len
-    }
-
-    async fn response_len(&self) -> usize {
-        self.inner.response_len
-    }
-
-    async fn status_code(&self) -> u16 {
-        self.inner.status_code
-    }
-
-    async fn status_msg(&self) -> &str {
-        &self.inner.status_msg
-    }
-
-    async fn username(&self) -> &str {
-        &self.inner.username
-    }
-
-    async fn password(&self) -> &str {
-        &self.inner.password
-    }
-
-    async fn cookie(&self) -> &str {
-        &self.inner.cookie
-    }
-
-    async fn content_encoding(&self) -> &str {
-        &self.inner.content_encoding
-    }
-
-    async fn content_type(&self) -> &str {
-        &self.inner.content_type
-    }
-
-    async fn cache_control(&self) -> &str {
-        &self.inner.cache_control
-    }
-
-    async fn db_name(&self) -> &str {
-        &self.inner.db_name
-    }
-
-    async fn rule_id(&self) -> u32 {
-        self.inner.rule_id
-    }
-
-    async fn matched_to(&self) -> &str {
-        &self.inner.matched_to
-    }
-
-    async fn cluster_id(&self) -> usize {
-        self.inner.cluster_id
-    }
-
-    async fn attack_kind(&self) -> &str {
-        &self.inner.attack_kind
-    }
-
-    async fn confidence(&self) -> f32 {
-        self.inner.confidence
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::HttpThreat> for HttpThreat {
-    fn from(inner: database::HttpThreat) -> Self {
-        Self { inner }
-    }
-}
-
-struct RdpBruteForce {
-    inner: database::RdpBruteForce,
-}
-
-#[Object]
-impl RdpBruteForce {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn dst_addrs(&self) -> Vec<String> {
-        self.inner
-            .dst_addrs
-            .iter()
-            .map(ToString::to_string)
-            .collect()
-    }
-
-    async fn start_time(&self) -> DateTime<Utc> {
-        self.inner.start_time
-    }
-
-    async fn last_time(&self) -> DateTime<Utc> {
-        self.inner.last_time
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::RdpBruteForce> for RdpBruteForce {
-    fn from(inner: database::RdpBruteForce) -> Self {
-        Self { inner }
-    }
-}
-
-struct RepeatedHttpSessions {
-    inner: database::RepeatedHttpSessions,
-}
-
-#[Object]
-impl RepeatedHttpSessions {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::RepeatedHttpSessions> for RepeatedHttpSessions {
-    fn from(inner: database::RepeatedHttpSessions) -> Self {
-        Self { inner }
-    }
-}
-
-struct TorConnection {
-    inner: database::TorConnection,
-}
-
-#[Object]
-impl TorConnection {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn session_end_time(&self) -> DateTime<Utc> {
-        self.inner.session_end_time
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn host(&self) -> &str {
-        &self.inner.host
-    }
-
-    async fn method(&self) -> &str {
-        &self.inner.method
-    }
-
-    async fn uri(&self) -> &str {
-        &self.inner.uri
-    }
-
-    async fn referer(&self) -> &str {
-        &self.inner.referrer
-    }
-
-    async fn version(&self) -> &str {
-        &self.inner.version
-    }
-
-    async fn user_agent(&self) -> &str {
-        &self.inner.user_agent
-    }
-
-    async fn request_len(&self) -> usize {
-        self.inner.request_len
-    }
-
-    async fn response_len(&self) -> usize {
-        self.inner.response_len
-    }
-
-    async fn status_code(&self) -> u16 {
-        self.inner.status_code
-    }
-
-    async fn status_msg(&self) -> &str {
-        &self.inner.status_msg
-    }
-
-    async fn username(&self) -> &str {
-        &self.inner.username
-    }
-
-    async fn password(&self) -> &str {
-        &self.inner.password
-    }
-
-    async fn cookie(&self) -> &str {
-        &self.inner.cookie
-    }
-
-    async fn content_encoding(&self) -> &str {
-        &self.inner.content_encoding
-    }
-
-    async fn content_type(&self) -> &str {
-        &self.inner.content_type
-    }
-
-    async fn cache_control(&self) -> &str {
-        &self.inner.cache_control
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::TorConnection> for TorConnection {
-    fn from(inner: database::TorConnection) -> Self {
-        Self { inner }
-    }
-}
-
-struct DomainGenerationAlgorithm {
-    inner: database::DomainGenerationAlgorithm,
-}
-
-#[Object]
-impl DomainGenerationAlgorithm {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn host(&self) -> &str {
-        &self.inner.host
-    }
-
-    async fn method(&self) -> &str {
-        &self.inner.method
-    }
-
-    async fn uri(&self) -> &str {
-        &self.inner.uri
-    }
-
-    async fn referer(&self) -> &str {
-        &self.inner.referer
-    }
-
-    async fn version(&self) -> &str {
-        &self.inner.version
-    }
-
-    async fn user_agent(&self) -> &str {
-        &self.inner.user_agent
-    }
-
-    async fn request_len(&self) -> usize {
-        self.inner.request_len
-    }
-
-    async fn response_len(&self) -> usize {
-        self.inner.response_len
-    }
-
-    async fn status_code(&self) -> u16 {
-        self.inner.status_code
-    }
-
-    async fn status_msg(&self) -> &str {
-        &self.inner.status_msg
-    }
-
-    async fn username(&self) -> &str {
-        &self.inner.username
-    }
-
-    async fn password(&self) -> &str {
-        &self.inner.password
-    }
-
-    async fn cookie(&self) -> &str {
-        &self.inner.cookie
-    }
-
-    async fn content_encoding(&self) -> &str {
-        &self.inner.content_encoding
-    }
-
-    async fn content_type(&self) -> &str {
-        &self.inner.content_type
-    }
-
-    async fn cache_control(&self) -> &str {
-        &self.inner.cache_control
-    }
-
-    async fn confidence(&self) -> f32 {
-        self.inner.confidence
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::DomainGenerationAlgorithm> for DomainGenerationAlgorithm {
-    fn from(inner: database::DomainGenerationAlgorithm) -> Self {
-        Self { inner }
-    }
-}
-
-struct FtpBruteForce {
-    inner: database::FtpBruteForce,
-}
-
-#[Object]
-impl FtpBruteForce {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn user_list(&self) -> Vec<String> {
-        self.inner.user_list.clone()
-    }
-
-    async fn start_time(&self) -> DateTime<Utc> {
-        self.inner.start_time
-    }
-
-    async fn last_time(&self) -> DateTime<Utc> {
-        self.inner.last_time
-    }
-
-    async fn is_internal(&self) -> bool {
-        self.inner.is_internal
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::FtpBruteForce> for FtpBruteForce {
-    fn from(inner: database::FtpBruteForce) -> Self {
-        Self { inner }
-    }
-}
-
-struct FtpPlainText {
-    inner: database::FtpPlainText,
-}
-
-#[Object]
-impl FtpPlainText {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn user(&self) -> &str {
-        &self.inner.user
-    }
-
-    async fn password(&self) -> &str {
-        &self.inner.password
-    }
-
-    async fn command(&self) -> &str {
-        &self.inner.command
-    }
-
-    async fn reply_code(&self) -> &str {
-        &self.inner.reply_code
-    }
-
-    async fn reply_msg(&self) -> &str {
-        &self.inner.reply_msg
-    }
-
-    async fn data_passive(&self) -> bool {
-        self.inner.data_passive
-    }
-
-    async fn data_orig_addr(&self) -> String {
-        self.inner.data_orig_addr.to_string()
-    }
-
-    async fn data_resp_addr(&self) -> String {
-        self.inner.data_resp_addr.to_string()
-    }
-
-    async fn data_resp_port(&self) -> u16 {
-        self.inner.data_resp_port
-    }
-
-    async fn file(&self) -> &str {
-        &self.inner.file
-    }
-
-    async fn file_size(&self) -> u64 {
-        self.inner.file_size
-    }
-
-    async fn file_id(&self) -> &str {
-        &self.inner.file_id
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::FtpPlainText> for FtpPlainText {
-    fn from(inner: database::FtpPlainText) -> Self {
-        Self { inner }
-    }
-}
-
-struct PortScan {
-    inner: database::PortScan,
-}
-
-#[Object]
-impl PortScan {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_ports(&self) -> Vec<u16> {
-        self.inner.dst_ports.clone()
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn start_time(&self) -> DateTime<Utc> {
-        self.inner.start_time
-    }
-
-    async fn last_time(&self) -> DateTime<Utc> {
-        self.inner.last_time
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::PortScan> for PortScan {
-    fn from(inner: database::PortScan) -> Self {
-        Self { inner }
-    }
-}
-
-struct MultiHostPortScan {
-    inner: database::MultiHostPortScan,
-}
-
-#[Object]
-impl MultiHostPortScan {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn dst_addrs(&self) -> Vec<String> {
-        self.inner
-            .dst_addrs
-            .iter()
-            .map(ToString::to_string)
-            .collect()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(
-            ctx,
-            *self
-                .inner
-                .dst_addrs
-                .get(0)
-                .expect("has value with internal network"),
-        )
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(
-            &map,
-            *self
-                .inner
-                .dst_addrs
-                .get(0)
-                .expect("has value with internal network"),
-        )
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(
-            &map,
-            *self
-                .inner
-                .dst_addrs
-                .get(0)
-                .expect("has value with internal network"),
-        )
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn start_time(&self) -> DateTime<Utc> {
-        self.inner.start_time
-    }
-
-    async fn last_time(&self) -> DateTime<Utc> {
-        self.inner.last_time
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::MultiHostPortScan> for MultiHostPortScan {
-    fn from(inner: database::MultiHostPortScan) -> Self {
-        Self { inner }
-    }
-}
-
-struct ExternalDdos {
-    inner: database::ExternalDdos,
-}
-
-#[Object]
-impl ExternalDdos {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn src_addrs(&self) -> Vec<String> {
-        self.inner
-            .src_addrs
-            .iter()
-            .map(ToString::to_string)
-            .collect()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(
-            ctx,
-            *self
-                .inner
-                .src_addrs
-                .get(0)
-                .expect("has value with internal network"),
-        )
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(
-            &map,
-            *self
-                .inner
-                .src_addrs
-                .get(0)
-                .expect("has value with internal network"),
-        )
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(
-            &map,
-            *self
-                .inner
-                .src_addrs
-                .get(0)
-                .expect("has value with internal network"),
-        )
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn start_time(&self) -> DateTime<Utc> {
-        self.inner.start_time
-    }
-
-    async fn last_time(&self) -> DateTime<Utc> {
-        self.inner.last_time
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::ExternalDdos> for ExternalDdos {
-    fn from(inner: database::ExternalDdos) -> Self {
-        Self { inner }
-    }
-}
-
-struct NonBrowser {
-    inner: database::NonBrowser,
-}
-
-#[Object]
-impl NonBrowser {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn host(&self) -> &str {
-        &self.inner.host
-    }
-
-    async fn method(&self) -> &str {
-        &self.inner.method
-    }
-
-    async fn uri(&self) -> &str {
-        &self.inner.uri
-    }
-
-    async fn referer(&self) -> &str {
-        &self.inner.referrer
-    }
-
-    async fn version(&self) -> &str {
-        &self.inner.version
-    }
-
-    async fn user_agent(&self) -> &str {
-        &self.inner.user_agent
-    }
-
-    async fn request_len(&self) -> usize {
-        self.inner.request_len
-    }
-
-    async fn response_len(&self) -> usize {
-        self.inner.response_len
-    }
-
-    async fn status_code(&self) -> u16 {
-        self.inner.status_code
-    }
-
-    async fn status_msg(&self) -> &str {
-        &self.inner.status_msg
-    }
-
-    async fn username(&self) -> &str {
-        &self.inner.username
-    }
-
-    async fn password(&self) -> &str {
-        &self.inner.password
-    }
-
-    async fn cookie(&self) -> &str {
-        &self.inner.cookie
-    }
-
-    async fn content_encoding(&self) -> &str {
-        &self.inner.content_encoding
-    }
-
-    async fn content_type(&self) -> &str {
-        &self.inner.content_type
-    }
-
-    async fn cache_control(&self) -> &str {
-        &self.inner.cache_control
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::NonBrowser> for NonBrowser {
-    fn from(inner: database::NonBrowser) -> Self {
-        Self { inner }
-    }
-}
-
-struct LdapBruteForce {
-    inner: database::LdapBruteForce,
-}
-
-#[Object]
-impl LdapBruteForce {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn user_pw_list(&self) -> Vec<String> {
-        self.inner
-            .user_pw_list
-            .iter()
-            .map(|(user, pw)| format!("{user}/{pw}"))
-            .collect()
-    }
-
-    async fn start_time(&self) -> DateTime<Utc> {
-        self.inner.start_time
-    }
-
-    async fn last_time(&self) -> DateTime<Utc> {
-        self.inner.last_time
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::LdapBruteForce> for LdapBruteForce {
-    fn from(inner: database::LdapBruteForce) -> Self {
-        Self { inner }
-    }
-}
-
-struct LdapPlainText {
-    inner: database::LdapPlainText,
-}
-
-#[Object]
-impl LdapPlainText {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn message_id(&self) -> u32 {
-        self.inner.message_id
-    }
-
-    async fn version(&self) -> u8 {
-        self.inner.version
-    }
-
-    async fn opcode(&self) -> Vec<String> {
-        self.inner.opcode.clone()
-    }
-
-    async fn result(&self) -> Vec<String> {
-        self.inner.result.clone()
-    }
-
-    async fn diagnostic_message(&self) -> Vec<String> {
-        self.inner.diagnostic_message.clone()
-    }
-
-    async fn object(&self) -> Vec<String> {
-        self.inner.object.clone()
-    }
-
-    async fn argument(&self) -> Vec<String> {
-        self.inner.argument.clone()
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::LdapPlainText> for LdapPlainText {
-    fn from(inner: database::LdapPlainText) -> Self {
-        Self { inner }
-    }
-}
-
-struct CryptocurrencyMiningPool {
-    inner: database::CryptocurrencyMiningPool,
-}
-
-#[Object]
-impl CryptocurrencyMiningPool {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn query(&self) -> &str {
-        &self.inner.query
-    }
-
-    async fn answer(&self) -> Vec<String> {
-        self.inner.answer.clone()
-    }
-
-    async fn trans_id(&self) -> u16 {
-        self.inner.trans_id
-    }
-
-    async fn rtt(&self) -> i64 {
-        self.inner.rtt
-    }
-
-    async fn qclass(&self) -> u16 {
-        self.inner.qclass
-    }
-
-    async fn qtype(&self) -> u16 {
-        self.inner.qtype
-    }
-
-    async fn rcode(&self) -> u16 {
-        self.inner.rcode
-    }
-
-    async fn aa_flag(&self) -> bool {
-        self.inner.aa_flag
-    }
-
-    async fn tc_flag(&self) -> bool {
-        self.inner.tc_flag
-    }
-
-    async fn rd_flag(&self) -> bool {
-        self.inner.rd_flag
-    }
-
-    async fn ra_flag(&self) -> bool {
-        self.inner.ra_flag
-    }
-
-    async fn ttl(&self) -> Vec<i32> {
-        self.inner.ttl.clone()
-    }
-
-    async fn coins(&self) -> Vec<String> {
-        self.inner.coins.clone()
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::CryptocurrencyMiningPool> for CryptocurrencyMiningPool {
-    fn from(inner: database::CryptocurrencyMiningPool) -> Self {
-        Self { inner }
-    }
-}
-
-struct BlockListConn {
-    inner: database::BlockListConn,
-}
-
-#[Object]
-impl BlockListConn {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn duration(&self) -> i64 {
-        self.inner.duration
-    }
-
-    async fn service(&self) -> String {
-        self.inner.service.clone()
-    }
-
-    async fn orig_bytes(&self) -> u64 {
-        self.inner.orig_bytes
-    }
-
-    async fn resp_bytes(&self) -> u64 {
-        self.inner.resp_bytes
-    }
-
-    async fn orig_pkts(&self) -> u64 {
-        self.inner.orig_pkts
-    }
-
-    async fn resp_pkts(&self) -> u64 {
-        self.inner.resp_pkts
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::BlockListConn> for BlockListConn {
-    fn from(inner: database::BlockListConn) -> Self {
-        Self { inner }
-    }
-}
-
-struct BlockListDns {
-    inner: database::BlockListDns,
-}
-
-#[Object]
-impl BlockListDns {
-    async fn time(&self) -> DateTime<Utc> {
-        self.inner.time
-    }
-
-    async fn source(&self) -> &str {
-        &self.inner.source
-    }
-
-    async fn src_addr(&self) -> String {
-        self.inner.src_addr.to_string()
-    }
-
-    /// The two-letter country code of the source IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn src_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.src_addr)
-    }
-
-    async fn src_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.src_addr)
-    }
-
-    async fn src_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.src_addr)
-    }
-
-    async fn src_port(&self) -> u16 {
-        self.inner.src_port
-    }
-
-    async fn dst_addr(&self) -> String {
-        self.inner.dst_addr.to_string()
-    }
-
-    /// The two-letter country code of the destination IP address. `"XX"` if the
-    /// location of the address is not known, and `"ZZ"` if the location
-    /// database is unavailable.
-    async fn dst_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.dst_addr)
-    }
-
-    async fn dst_customer(&self, ctx: &Context<'_>) -> Result<Option<Customer>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.customer_map();
-        find_ip_customer(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_network(&self, ctx: &Context<'_>) -> Result<Option<Network>> {
-        let store = crate::graphql::get_store(ctx).await?;
-        let map = store.network_map();
-        find_ip_network(&map, self.inner.dst_addr)
-    }
-
-    async fn dst_port(&self) -> u16 {
-        self.inner.dst_port
-    }
-
-    async fn proto(&self) -> u8 {
-        self.inner.proto
-    }
-
-    async fn last_time(&self) -> i64 {
-        self.inner.last_time
-    }
-
-    async fn query(&self) -> &str {
-        &self.inner.query
-    }
-
-    async fn answer(&self) -> Vec<String> {
-        self.inner.answer.clone()
-    }
-
-    async fn trans_id(&self) -> u16 {
-        self.inner.trans_id
-    }
-
-    async fn rtt(&self) -> i64 {
-        self.inner.rtt
-    }
-
-    async fn qclass(&self) -> u16 {
-        self.inner.qclass
-    }
-
-    async fn qtype(&self) -> u16 {
-        self.inner.qtype
-    }
-
-    async fn rcode(&self) -> u16 {
-        self.inner.rcode
-    }
-
-    async fn aa_flag(&self) -> bool {
-        self.inner.aa_flag
-    }
-
-    async fn tc_flag(&self) -> bool {
-        self.inner.tc_flag
-    }
-
-    async fn rd_flag(&self) -> bool {
-        self.inner.rd_flag
-    }
-
-    async fn ra_flag(&self) -> bool {
-        self.inner.ra_flag
-    }
-
-    async fn ttl(&self) -> Vec<i32> {
-        self.inner.ttl.clone()
-    }
-
-    async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
-        self.inner
-            .triage_scores
-            .as_ref()
-            .map(|scores| scores.iter().map(Into::into).collect::<Vec<TriageScore>>())
-    }
-}
-
-impl From<database::BlockListDns> for BlockListDns {
-    fn from(inner: database::BlockListDns) -> Self {
-        Self { inner }
-    }
 }
 
 struct EventTotalCount {
